REFACTOR: only pass SampleInfoIterator when needed
We don't need to create a SampleInfoIterator until we're actually
passing data to pgx, until that point a regular slice will do.
